Create class for stemmer
