+++ src/main/java/ru/cs/eatright/parsing/RequestProcessor.java
@@ -6,10 +6,10 @@ import edu.stanford.nlp.pipeline.StanfordCoreNLP;
 import edu.stanford.nlp.util.CoreMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.tartarus.snowball.ext.RussianStemmer;
 import ru.cs.eatright.nlp.Chunker;
 import ru.cs.eatright.nlp.Phrase;
 import ru.cs.eatright.nlp.RusPosAnnotator;
+import ru.cs.eatright.nlp.Stemmer;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -24,7 +24,7 @@ public class RequestProcessor {
     private final Tokenizer tokenizer;
     private StanfordCoreNLP pipeline = null;
     private Chunker chunker = new Chunker();
-    private final RussianStemmer stemmer = new RussianStemmer();
+    private Stemmer stemmer = new Stemmer();
 
     public RequestProcessor(Tokenizer tokenizer) {
         this.tokenizer = tokenizer;
@@ -49,34 +49,17 @@ public class RequestProcessor {
     private List<Query> convertRequest2StemmedQuery(String request) throws IOException {
         List<Token> tokens = tokenizer.tokenize(request);
         String cleanedText = cleanText(tokens);
-        List<Phrase> nounPhrases = extractNounPhrases(cleanedText);
+        List<CoreMap> sentences = getSentenceAnnotations(cleanedText);
+        List<Phrase> nounPhrases = getPhrases(sentences);
         return getQueries(nounPhrases);
     }
 
     private List<Query> getQueries(List<Phrase> nounPhrases) {
         List<Query> queries = new ArrayList<>();
         for (Phrase nounPhrase : nounPhrases) {
-            queries.add(new Query(nounPhrase.getPhraseWords(), getStemmedWords(nounPhrase.getPhraseWords())));
+            queries.add(new Query(nounPhrase.getPhraseWords(), stemmer.getStemmedWords(nounPhrase.getPhraseWords())));
         }
         return queries;
-    }
-
-    private List<String> getStemmedWords(List<String> phraseWords) {
-        List<String> stemmedPhraseWords = new ArrayList<>();
-        for (String phraseWord : phraseWords) {
-            if (phraseWord.length() >= 5) {
-                stemmedPhraseWords.add(stem(phraseWord));
-            } else {
-                stemmedPhraseWords.add(phraseWord);
-            }
-        }
-        return stemmedPhraseWords;
-    }
-
-    private String stem(String phraseWord) {
-        stemmer.setCurrent(phraseWord);
-        stemmer.stem();
-        return stemmer.getCurrent();
     }
 
     private String cleanText(List<Token> tokens) {
@@ -99,20 +82,16 @@ public class RequestProcessor {
         return filteredTokens;
     }
 
-    private List<Phrase> extractNounPhrases(String text) {
-        List<CoreMap> sentences = getSentenceAnnotations(text);
-
-        if (sentences == null || sentences.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        return getPhrases(sentences);
-    }
-
     private List<CoreMap> getSentenceAnnotations(String text) {
         Annotation annotation = new Annotation(text);
         pipeline.annotate(annotation);
-        return annotation.get(CoreAnnotations.SentencesAnnotation.class);
+        List<CoreMap> sentencesAnnotations = annotation.get(CoreAnnotations.SentencesAnnotation.class);
+
+        if (sentencesAnnotations == null || sentencesAnnotations.isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        return sentencesAnnotations;
     }
 
     private List<Phrase> getPhrases(List<CoreMap> sentences) {
